Fixed duplicate stream error
The listener side of the connection should not create a stream, it should instead wait for a stream.
The dialer side will send an empty sctp message upon connection to make the connection visible for the dialer side.
Without the empty message, the listener side would not be able to send messages before receiving a message from the dialer side.