Error are not generic anymore
Having generics on error types was too complex. Now, there is a predefined set of errors that implementations can choose from.
Some common errors are propagated to upper type like TimeoutError which is present everywhere.
Conversions are a bit rougher because we have to do all the propagation for these types exists on both sides of the conversion. However, users can always rely on the first level of error enum, making error handling simpler.