Public key based signaling signature
ED25519 is used to sign the public keys on the ephemeral diffie hellman protocol.
The static key is used to generate a shared secret that will be the channel name provided to the signaling server, so that the peer can find each other. This static key generated shared secret is not actually used in any encryption.